requirements.txt
bert_score==0.3.13
json_repair==0.48.0
lfss==0.13.3
numpy==2.3.4
pandas==2.3.3
Pillow==12.0.0
PyYAML==6.0.3
scikit_learn==1.7.2
torch==2.7.0
tqdm==4.67.1
transformers==4.56.2
ujson==5.11.0
openai==2.6.1